Restructure affiliate code
