+++ affiliate/package.json
