+++ affiliate/src/productCard/Edit.jsx
@@ -22,8 +22,8 @@ import {
 } from '@wordpress/element';
 import PropTypes from 'prop-types';
 
+import ProductSearchModal from '../ProductSearchModal';
 import ProductCard from './ProductCard';
-import ProductSearchModal from './ProductSearchModal';
 import { AttributesType } from './propTypes';
 
 const Edit = ({ attributes, setAttributes }) => {
