+++ affiliate/src/productCard/propTypes.js
